Implemented support for uploading files to S3 buckets

+++ GuiStack/Repositories/IS3Repository.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
@@ -21,6 +22,7 @@ namespace GuiStack.Repositories
         Task<IEnumerable<S3Object>> GetObjectsAsync(string bucketName);
         Task<Amazon.S3.Model.GetObjectResponse> GetObjectAsync(string bucketName, string objectName);
         Task RenameObjectAsync(string bucketName, string objectName, string newName);
+        Task UploadFile(string bucketName, string filename, Stream stream);
     }
 
     public class S3Repository : IS3Repository
@@ -38,6 +40,9 @@ namespace GuiStack.Repositories
 
         public async Task CreateBucketAsync(string bucketName)
         {
+            if(bucketName == null)
+                throw new ArgumentNullException(nameof(bucketName));
+
             using var s3 = authenticator.Authenticate();
             var response = await s3.PutBucketAsync(new Amazon.S3.Model.PutBucketRequest() {
                 BucketName = bucketName,
@@ -50,6 +55,9 @@ namespace GuiStack.Repositories
 
         public async Task DeleteBucketAsync(string bucketName)
         {
+            if(bucketName == null)
+                throw new ArgumentNullException(nameof(bucketName));
+
             using var s3 = authenticator.Authenticate();
             var response = await s3.DeleteBucketAsync(bucketName);
 
@@ -58,6 +66,12 @@ namespace GuiStack.Repositories
 
         public async Task DeleteObjectAsync(string bucketName, string objectName)
         {
+            if(bucketName == null)
+                throw new ArgumentNullException(nameof(bucketName));
+
+            if(objectName == null)
+                throw new ArgumentNullException(nameof(objectName));
+
             using var s3 = authenticator.Authenticate();
             var response = await s3.DeleteObjectAsync(bucketName, objectName);
 
@@ -134,6 +148,98 @@ namespace GuiStack.Repositories
 
             var deleteResponse = await s3.DeleteObjectAsync(bucketName, objectName);
             deleteResponse.ThrowIfUnsuccessful("S3 delete");
+        }
+
+        public async Task UploadFile(string bucketName, string filename, Stream stream)
+        {
+            if(bucketName == null)
+                throw new ArgumentNullException(nameof(bucketName));
+
+            if(filename == null)
+                throw new ArgumentNullException(nameof(filename));
+
+            if(stream == null)
+                throw new ArgumentNullException(nameof(stream));
+
+            if(!stream.CanRead)
+                throw new ArgumentException("File upload stream must be readable", nameof(stream));
+
+            long fileSize;
+
+            try
+            {
+                fileSize = stream.Length;
+            }
+            catch(Exception ex)
+            {
+                throw new ArgumentException("Failed to get the total upload file size", ex);
+            }
+
+            using var s3 = authenticator.Authenticate();
+
+            var initiateUploadResponse = await s3.InitiateMultipartUploadAsync(new Amazon.S3.Model.InitiateMultipartUploadRequest() {
+                BucketName = bucketName,
+                Key = filename
+            });
+
+            initiateUploadResponse.ThrowIfUnsuccessful("S3");
+
+            // 5 MB is the minimum allowed chunk size for multipart uploads
+            const int ChunkSize = 5 * 1024 * 1024;
+
+            long bytesWritten = 0;
+            int partNumber = 0;
+
+            string uploadId = initiateUploadResponse.UploadId;
+
+            try
+            {
+                var uploadResponses = new List<Amazon.S3.Model.UploadPartResponse>();
+
+                while(bytesWritten < fileSize)
+                {
+                    long remaining = fileSize - bytesWritten;
+                    bool isLastPart = remaining <= ChunkSize;
+                    int partSize = isLastPart ? (int)remaining : ChunkSize;
+
+                    partNumber++;
+
+                    var response = await s3.UploadPartAsync(new Amazon.S3.Model.UploadPartRequest() {
+                        UploadId = uploadId,
+                        BucketName = bucketName,
+                        Key = filename,
+                        InputStream = stream,
+                        PartSize = partSize,
+                        PartNumber = partNumber,
+                        IsLastPart = isLastPart,
+                    });
+
+                    response.ThrowIfUnsuccessful("S3");
+
+                    bytesWritten += partSize;
+                    uploadResponses.Add(response);
+
+                    if(isLastPart)
+                        break;
+                }
+
+                var completeUploadRequest = new Amazon.S3.Model.CompleteMultipartUploadRequest() {
+                    UploadId = uploadId,
+                    BucketName = bucketName,
+                    Key = filename,
+                };
+
+                completeUploadRequest.AddPartETags(uploadResponses);
+
+                (await s3.CompleteMultipartUploadAsync(completeUploadRequest)).ThrowIfUnsuccessful("S3");
+            }
+            catch
+            {
+                // Deliberately ignore AWS API response, since we don't care if an error occurred here
+                await s3.AbortMultipartUploadAsync(bucketName, filename, uploadId);
+
+                throw;
+            }
         }
     }
 }
